@@ -41,7 +41,11 @@ def format_query(stats: list[str], rarities: list[int]) -> str:
 
 
 def stat_wheeling(
-    stat_mins: SetMinimums, available_slots: list[str], allowed_rarities: list[int], min_lv: int, max_lv: int
+    stat_mins: SetMinimums,
+    available_slots: list[str],
+    allowed_rarities: list[int],
+    min_lv: int,
+    max_lv: int,
 ) -> dict[str, list[tuple[int]]]:
     ret: dict[str, list[tuple[int]]] = {}
     if not available_slots:
@@ -58,7 +62,10 @@ def stat_wheeling(
     with conn:
         for s in set(available_slots):
             ret[s] = list(
-                cursor.execute(format_query(cols, allowed_rarities), {"min_lv": min_lv, "max_lv": max_lv, "position": s})
+                cursor.execute(
+                    format_query(cols, allowed_rarities),
+                    {"min_lv": min_lv, "max_lv": max_lv, "position": s},
+                )
             )
 
     return ret
@@ -68,28 +75,44 @@ AllowedStatType = Literal["ap", "mp", "wp", "ra"]
 
 
 def all_valid_by_slot(
-    stat_mins: SetMinimums, available_slots: list[str], allowed_rarities: list[int], min_lv: int, max_lv: int
+    stat_mins: SetMinimums,
+    available_slots: list[str],
+    allowed_rarities: list[int],
+    min_lv: int,
+    max_lv: int,
 ) -> Iterator[tuple[str, tuple[int]]]:
     allowed: list[AllowedStatType] = ["ap", "mp", "wp", "ra"]
-    cols: list[AllowedStatType] = [s for s in allowed if getattr(stat_mins, s, DUMMY_MIN) > DUMMY_MIN]
+    cols: list[AllowedStatType] = [
+        s for s in allowed if getattr(stat_mins, s, DUMMY_MIN) > DUMMY_MIN
+    ]
     data = stat_wheeling(
-        stat_mins=stat_mins, available_slots=available_slots, allowed_rarities=allowed_rarities, min_lv=min_lv, max_lv=max_lv
+        stat_mins=stat_mins,
+        available_slots=available_slots,
+        allowed_rarities=allowed_rarities,
+        min_lv=min_lv,
+        max_lv=max_lv,
     )
 
     for combo in itertools.product(*[data[k] for k in available_slots]):
         elem_wise_sum = add.reduce(combo)
 
-        temp: dict[AllowedStatType, int] = {k: v for k, v in zip(cols, elem_wise_sum)}  # noqa: C416
+        temp: dict[AllowedStatType, int] = dict(zip(cols, elem_wise_sum, strict=False))
 
         s = Stats(**temp)
 
         if stat_mins.stats_met(s):
-            yield from zip(available_slots, combo)
+            yield from zip(available_slots, combo, strict=False)
 
 
 if __name__ == "__main__":
     """ This will take an impossibly long time to run for just a single scenario limited to only these stats """
     s: dict[str, set[tuple[int, ...]]] = defaultdict(set)
-    for slot, stats in all_valid_by_slot(SetMinimums(ap=12, mp=6, wp=0, ra=0), v1BuildSlotsOrder, [1, 2, 3, 4, 5, 6, 7], 1, 230):
+    for slot, stats in all_valid_by_slot(
+        SetMinimums(ap=12, mp=6, wp=0, ra=0),
+        v1BuildSlotsOrder,
+        [1, 2, 3, 4, 5, 6, 7],
+        1,
+        230,
+    ):
         s[slot].add(stats)
     print(s)  # noqa: T201
